Define wait for images to be exported in conftest
