+++ test/conftest.py
@@ -15,9 +15,14 @@
 
 # ruff: noqa: C408 dict() makes test data easier to read and write
 import os
-from collections.abc import Generator
+from functools import partial, update_wrapper
 from pathlib import Path
 from typing import Any
+
+from core.db.models import Image
+from core.db.queries import engine
+from sqlalchemy import cast, not_
+from sqlalchemy.orm import sessionmaker
 
 # Setting env variables before loading modules
 os.environ["PIXL_DB_HOST"] = "localhost"
@@ -30,9 +35,8 @@ import pytest
 import requests
 from pytest_pixl.dicom import generate_dicom_dataset
 from pytest_pixl.ftpserver import PixlFTPServer
-from pytest_pixl.helpers import run_subprocess
+from pytest_pixl.helpers import run_subprocess, wait_for_condition
 from pytest_pixl.plugin import FtpHostAddress
-from utils import wait_for_images_to_be_exported
 
 pytest_plugins = "pytest_pixl"
 
@@ -142,12 +146,50 @@ def _upload_dicom_instance(dicom_dir: Path, **kwargs: Any) -> None:
     _upload_to_vna(test_dcm_file)
 
 
+def wait_for_images_to_be_exported(
+    seconds_max: int,
+    seconds_interval: int,
+    seconds_condition_stays_true_for: int,
+    min_studies: int = 3,
+) -> None:
+    """
+    Query pixl DB to ensure that images have been processed and exported.
+    If they haven't within the time limit, raise a TimeoutError
+    """
+    studies: list[Image] = []
+
+    def at_least_n_studies_exported(n_studies: int) -> bool:
+        nonlocal studies
+
+        PixlSession = sessionmaker(engine)
+        with PixlSession() as session:
+            studies = cast(
+                list[Image],
+                session.query(Image).filter(not_(Image.exported_at.is_(None))).all(),
+            )
+        return len(studies) >= n_studies
+
+    condition = partial(at_least_n_studies_exported, min_studies)
+    update_wrapper(condition, at_least_n_studies_exported)
+
+    def list_studies() -> str:
+        return f"Expecting at least {min_studies} studies.\nexported studies: {studies}"
+
+    wait_for_condition(
+        condition,
+        seconds_max=seconds_max,
+        seconds_interval=seconds_interval,
+        progress_string_fn=list_studies,
+        seconds_condition_stays_true_for=seconds_condition_stays_true_for,
+    )
+
+
 @pytest.fixture(scope="session")
 def _setup_pixl_cli(ftps_server: PixlFTPServer, _populate_vna: None) -> None:
     """Run pixl populate/start. Cleanup intermediate export dir on exit."""
     run_subprocess(["pixl", "populate", str(RESOURCES_OMOP_DIR.absolute())], TEST_DIR)
     # poll here for two minutes to check for imaging to be processed, printing progress
-    wait_for_stable_orthanc_anon(181, 5, 15, min_instances=3)
+    wait_for_images_to_be_exported(181, 5, 15)
 
 
 @pytest.fixture(scope="session")
@@ -155,7 +197,7 @@ def _setup_pixl_cli_dicomweb(_populate_vna: None) -> None:
     """Run pixl populate/start. Cleanup intermediate export dir on exit."""
     run_subprocess(["pixl", "populate", str(RESOURCES_OMOP_DICOMWEB_DIR.absolute())], TEST_DIR)
     # poll here for two minutes to check for imaging to be processed, printing progress
-    wait_for_stable_orthanc_anon(181, 5, 15, min_instances=3)
+    wait_for_images_to_be_exported(181, 5, 15)
 
 
 @pytest.fixture(scope="session")
